using connection info for test
